Split View to View and Model
