+++ dl_markup/view.py
@@ -1,67 +1,45 @@
-from PyQt5.QtWidgets import QLabel
 from PyQt5.QtWidgets import QListView
 from PyQt5.QtWidgets import QVBoxLayout
 from PyQt5.QtWidgets import QHBoxLayout
 from PyQt5.QtWidgets import QWidget
 from PyQt5.QtWidgets import QGraphicsView
 from PyQt5.QtWidgets import QToolBar
-from PyQt5.QtWidgets import QFileDialog
 from PyQt5.QtWidgets import QMainWindow
-import os
 
-from .list_model import ListModel
+from .model import Model
 
 
 class View(QMainWindow):
     """
     Application view
     """
-    def __init__(self, scene):
+    def __init__(self, model):
         super().__init__()
-        self.scene = scene
-        self.listModel = ListModel()
-        self.fileList = QListView()
-        self.fileList.setModel(self.listModel)
         self.setWindowTitle("DL Markup")
-        self.inputDirectory = QLabel('.')
-        self.outputDirectory = QLabel('.')
 
         mainLayout = QVBoxLayout()
         centralWidget = QWidget(self)
         self.setCentralWidget(centralWidget)
         centralWidget.setLayout(mainLayout)
-        mainLayout.addWidget(self.inputDirectory)
-        mainLayout.addWidget(self.outputDirectory)
+
+        mainLayout.addWidget(model.inputDirectory)
+        mainLayout.addWidget(model.outputDirectory)
 
         layout = QHBoxLayout()
         mainLayout.addLayout(layout)
+
+        self.fileList = QListView()
+        self.fileList.setModel(model.listModel)
         layout.addWidget(self.fileList)
-        layout.addWidget(QGraphicsView(scene))
 
-        self._createToolbar()
+        layout.addWidget(QGraphicsView(model.scene))
 
-    def _createToolbar(self):
+        self._createToolbar(model)
+
+    def _createToolbar(self, model):
         tools = QToolBar()
         self.addToolBar(tools)
-        tools.addAction('Select input directory', self._selectInputDirectory)
-        tools.addAction('Select output directory', self._selectOutputDirectory)
-        tools.addAction('Open', self._open)
-        tools.addAction('Save', self._save)
-
-    def _selectInputDirectory(self):
-        self.inputDirectory.setText(QFileDialog.getExistingDirectory())
-        self._updateFileList()
-
-    def _selectOutputDirectory(self):
-        self.outputDirectory.setText(QFileDialog.getExistingDirectory())
-
-    def _open(self):
-        pass
-
-    def _save(self):
-        pass
-
-    def _updateFileList(self):
-        files = os.listdir(self.inputDirectory.text())
-        print("Updating file list:", files)
-        self.listModel.setItems(files)
+        tools.addAction('Select input directory', model.selectInputDirectory)
+        tools.addAction('Select output directory', model.selectOutputDirectory)
+        tools.addAction('Open', model.open)
+        tools.addAction('Save', model.save)
